Clipers/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.6.0
yt-dlp==2023.11.16
requests==2.31.0
python-multipart==0.0.6
librosa==0.10.1
pydub==0.25.1
numpy>=1.26.0
scipy>=1.11.0
matplotlib>=3.7.0
soundfile==0.12.1
google-generativeai==0.7.1
python-dotenv==1.0.0
